layout mismatch cleanup
