+++ shame/src/frontend/rust_types/layout_traits.rs
@@ -136,9 +136,10 @@ use std::rc::Rc;
 /// [`StorageTexture`]: crate::StorageTexture
 ///
 pub trait GpuLayout {
-    /// TODO(chronicl)
+    /// Returns a [`TypeLayoutRecipe`] that describes how the type is laid out in memory.
     fn layout_recipe() -> TypeLayoutRecipe;
-    /// TODO(chronicl)
+
+    /// For `GpuSized` types, this returns the [`SizedType`] that describes the type's layout.
     fn layout_recipe_sized() -> SizedType
     where
         Self: GpuSized,
