+++ shame/src/frontend/rust_types/type_layout/compatible_with.rs
@@ -4,7 +4,11 @@ use crate::{
     any::layout::StructLayout,
     call_info,
     common::prettify::{set_color, UnwrapOrStr},
-    frontend::rust_types::type_layout::{display::LayoutInfo, ArrayLayout},
+    frontend::rust_types::type_layout::{
+        display::LayoutInfo,
+        eq::{try_find_mismatch, LayoutMismatch, StructMismatch, TopLevelMismatch},
+        ArrayLayout,
+    },
     ir::{ir_type::max_u64_po2_dividing, recording::Context},
     TypeLayout,
 };
@@ -24,7 +28,9 @@ use super::{recipe::TypeLayoutRecipe, Repr};
 /// Wgsl has only one representation of types - there is no choice between std140 and std430
 /// like in glsl - so to be representable in wgsl means that the type layout produced by
 /// the recipe is the same as the one produced by the same recipe but with all structs
-/// in the recipe using Repr::Wgsl, which is what shame calls wgsl's representation/layout algorithm.
+/// in the recipe using the Repr::Wgsl layout algorithm. Aditionally, all custom attributes used
+/// by the recipe need to be support by wgsl, which are only the struct field attributes
+/// `#[align(N)]` and `#[size(N)]` currently.
 pub struct TypeLayoutCompatibleWith<AddressSpace> {
     recipe: TypeLayoutRecipe,
     _phantom: std::marker::PhantomData<AddressSpace>,
@@ -46,7 +52,9 @@ impl<AS: AddressSpace> TypeLayoutCompatibleWith<AS> {
             AddressSpaceEnum::WgslStorage | AddressSpaceEnum::WgslUniform => {
                 // Wgsl has only one type representation: Repr::Wgsl, so the layout produced by the recipe
                 // is representable in wgsl iff the layout produced by the same recipe but with
-                // all structs in the recipe using Repr::Wgsl is the same.
+                // all structs in the recipe using Repr::Wgsl is the same and all custom attributes
+                // used by the recipe are supported by wgsl, which is checked in `TypeLayoutRecipe::layout`
+                // TODO(chronicl) line above
                 let recipe_unified = recipe.to_unified_repr(Repr::Wgsl);
                 let layout_unified = recipe_unified.layout();
                 if layout != layout_unified {
@@ -350,229 +358,4 @@ where
     writer.writeln_struct_end(f)?;
 
     Ok(())
-}
-
-/// Contains information about the layout mismatch between two `TypeLayout`s.
-///
-/// The type layouts are traversed depth-first and the first mismatch encountered
-/// is reported at its deepest level.
-///
-/// In case of nested structs this means that if the field `a` in
-/// ```
-/// struct A {
-///     a: struct B { ... }
-/// }
-/// struct AOther {
-///     a: struct BOther { ... }
-/// }
-/// ```
-/// mismatches, because `B` and `BOther` don't match, then the exact mismatch
-/// between `B` and `BOther` is reported and not the field mismatch of `a` in `A` and `AOther`.
-///
-/// Nested arrays are reported in two levels: `LayoutMismatch::TopLevel` contains the
-/// top level / outer most array layout and a  `TopLevelMismatch`, which contains the
-/// inner type layout where the mismatch is happening.
-/// For example if there is an array stride mismatch of the inner array of `Array<Array<f32x3>>`,
-/// then `LayoutMismatch::TopLevel` contains the layout of `Array<Array<f32x3>>` and
-/// a `TopLevelMismatch::ArrayStride` with the layout of `Array<f32x3>`.
-///
-/// A field of nested arrays in a struct is handled in the same way by
-/// `LayoutMismatch::Struct` containing the field index, which let's us access the outer
-/// most array, and a `TopLevelMismatch`, which let's us access the inner type layout
-/// where the mismatch is happening.
-#[derive(Debug, Clone)]
-pub enum LayoutMismatch {
-    TopLevel {
-        layout_left: TypeLayout,
-        layout_right: TypeLayout,
-        mismatch: TopLevelMismatch,
-    },
-    Struct {
-        struct_left: StructLayout,
-        struct_right: StructLayout,
-        mismatch: StructMismatch,
-    },
-}
-
-#[derive(Debug, Clone)]
-pub enum TopLevelMismatch {
-    Type,
-    ByteSize {
-        left: TypeLayout,
-        right: TypeLayout,
-    },
-    ArrayStride {
-        array_left: ArrayLayout,
-        array_right: ArrayLayout,
-    },
-}
-
-/// Field count is checked last.
-#[derive(Debug, Clone)]
-pub enum StructMismatch {
-    FieldName {
-        field_index: usize,
-    },
-    FieldLayout {
-        field_index: usize,
-        mismatch: TopLevelMismatch,
-    },
-    FieldOffset {
-        field_index: usize,
-    },
-    FieldCount,
-}
-
-/// Find the first depth first layout mismatch
-pub(crate) fn try_find_mismatch(layout1: &TypeLayout, layout2: &TypeLayout) -> Option<LayoutMismatch> {
-    use TypeLayout::*;
-
-    // First check if the kinds are the same type
-    match (&layout1, &layout2) {
-        (Vector(v1), Vector(v2)) => {
-            if v1.ty != v2.ty {
-                return Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1.clone(),
-                    layout_right: layout2.clone(),
-                    mismatch: TopLevelMismatch::Type,
-                });
-            }
-        }
-        (PackedVector(p1), PackedVector(p2)) => {
-            if p1.ty != p2.ty {
-                return Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1.clone(),
-                    layout_right: layout2.clone(),
-                    mismatch: TopLevelMismatch::Type,
-                });
-            }
-        }
-        (Matrix(m1), Matrix(m2)) => {
-            if m1.ty != m2.ty {
-                return Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1.clone(),
-                    layout_right: layout2.clone(),
-                    mismatch: TopLevelMismatch::Type,
-                });
-            }
-        }
-        (Array(a1), Array(a2)) => {
-            // Recursively check element types
-            match try_find_mismatch(&a1.element_ty, &a2.element_ty) {
-                // Update the top level layouts and propagate the LayoutMismatch
-                Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1,
-                    layout_right: layout2,
-                    mismatch,
-                }) => {
-                    return Some(LayoutMismatch::TopLevel {
-                        layout_left: layout1.clone(),
-                        layout_right: layout2.clone(),
-                        mismatch,
-                    });
-                }
-                // Struct mismatch, so it's not a top-level mismatch anymore
-                m @ Some(LayoutMismatch::Struct { .. }) => return m,
-                None => return None,
-            }
-
-            // Check array sizes match
-            if a1.len != a2.len {
-                return Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1.clone(),
-                    layout_right: layout2.clone(),
-                    mismatch: TopLevelMismatch::Type,
-                });
-            }
-
-            // Check array stride
-            if a1.byte_stride != a2.byte_stride {
-                return Some(LayoutMismatch::TopLevel {
-                    layout_left: layout1.clone(),
-                    layout_right: layout2.clone(),
-                    mismatch: TopLevelMismatch::ArrayStride {
-                        array_left: (**a1).clone(),
-                        array_right: (**a2).clone(),
-                    },
-                });
-            }
-        }
-        (Struct(s1), Struct(s2)) => {
-            return try_find_struct_mismatch(s1, s2);
-        }
-        // Different kinds entirely. Matching exhaustively, so that changes to TypeLayout lead us here.
-        (Vector(_) | PackedVector(_) | Matrix(_) | Array(_) | Struct(_), _) => {
-            return Some(LayoutMismatch::TopLevel {
-                layout_left: layout1.clone(),
-                layout_right: layout2.clone(),
-                mismatch: TopLevelMismatch::Type,
-            });
-        }
-    }
-
-    // Check byte size
-    if layout1.byte_size() != layout2.byte_size() {
-        return Some(LayoutMismatch::TopLevel {
-            layout_left: layout1.clone(),
-            layout_right: layout2.clone(),
-            mismatch: TopLevelMismatch::ByteSize {
-                left: layout1.clone(),
-                right: layout2.clone(),
-            },
-        });
-    }
-
-    None
-}
-
-fn try_find_struct_mismatch(struct1: &StructLayout, struct2: &StructLayout) -> Option<LayoutMismatch> {
-    for (field_index, (field1, field2)) in struct1.fields.iter().zip(struct2.fields.iter()).enumerate() {
-        // Order of checks is important here. We check in order
-        // - field name
-        // - field type, byte size and if the field is/contains a struct, recursively check its fields
-        // - field offset
-        if field1.name != field2.name {
-            return Some(LayoutMismatch::Struct {
-                struct_left: struct1.clone(),
-                struct_right: struct2.clone(),
-                mismatch: StructMismatch::FieldName { field_index },
-            });
-        }
-
-        // Recursively check field types
-        if let Some(inner_mismatch) = try_find_mismatch(&field1.ty, &field2.ty) {
-            match inner_mismatch {
-                // If it's a top-level mismatch, convert it to a field mismatch
-                LayoutMismatch::TopLevel { mismatch, .. } => {
-                    return Some(LayoutMismatch::Struct {
-                        struct_left: struct1.clone(),
-                        struct_right: struct2.clone(),
-                        mismatch: StructMismatch::FieldLayout { field_index, mismatch },
-                    });
-                }
-                // Pass through nested struct mismatches
-                struct_mismatch @ LayoutMismatch::Struct { .. } => return Some(struct_mismatch),
-            }
-        }
-
-        // Check field offset
-        if field1.rel_byte_offset != field2.rel_byte_offset {
-            return Some(LayoutMismatch::Struct {
-                struct_left: struct1.clone(),
-                struct_right: struct2.clone(),
-                mismatch: StructMismatch::FieldOffset { field_index },
-            });
-        }
-    }
-
-    // Check field count
-    if struct1.fields.len() != struct2.fields.len() {
-        return Some(LayoutMismatch::Struct {
-            struct_left: struct1.clone(),
-            struct_right: struct2.clone(),
-            mismatch: StructMismatch::FieldCount,
-        });
-    }
-
-    None
 }
